+++ dca/tl_content.php
@@ -10,6 +10,9 @@
  */
 
 
+//@todo check permisson
+
+
 /**
  * Dynamically add the permission check and parent table
  */