+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - add template support for Google Maps info windows <br>
   IMPORTANT: To override templates put it in the templates root folder! The AJAX request has no layout object and can not find the template in other folders.
 - runonce.php support
+- add full user / groups permission support
 
 ###Changed
 - change ajax request to new class FrontendAjax.php